Move `DowntimeList` to `Icinga\Module\Icingadb\Widget\ItemList`
